# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API BASE",
        "description": "Servicio API REST para la gestión y administración del sistema Automatizado de riego por goteo",
        "contact": {
            "name": "SETI",
            "email": "cirojas@cenicana.org"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/estate-irrigation-systems": {
            "get": {
                "tags": [
                    "EstateIrrigationSystems"
                ],
                "summary": "Get list of irrigation systems",
                "description": "Returns list of irrigation systems filter by its type of system and irrgation header \n     *      (filtro por tipo de sistema de riego o cabezal de riego)",
                "operationId": "getEstateIrrigationSystemList",
                "parameters": [
                    {
                        "name": "estate_id",
                        "in": "query",
                        "description": "The ID of the estate (hacienda)",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "irrigation_system_id",
                        "in": "query",
                        "description": "The ID of the type of system (tipo: goteo, aspersión, ..)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "irrigation_header_id",
                        "in": "query",
                        "description": "The ID of the irrigation header of the system (cabezal de riego)",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "EstateIrrigationSystems"
                ],
                "summary": "Store new estate irrigation system",
                "description": "Returns estate irrigation system data",
                "operationId": "storeEstateIrrigationSystem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EstateIrrigationSystem"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EstateIrrigationSystem"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/estate-irrigation-systems/{id}": {
            "get": {
                "tags": [
                    "EstateIrrigationSystems"
                ],
                "summary": "Get estate irrigation system information",
                "description": "Returns estate irrigation system data",
                "operationId": "getEstateIrrigationSystemById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Estate irrigation system id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/irrigation-systems": {
            "get": {
                "tags": [
                    "IrrigationSystems"
                ],
                "summary": "Get list of irrigation systems",
                "description": "Returns list of irrigation systems",
                "operationId": "getIrrigationSystemList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
        "/v1/irrigation-strategies": {
            "get": {
                "tags": [
                    "IrrigationEstrategies"
                ],
                "summary": "Get list of irrigation estrategies",
                "description": "Returns list of irrigation estrategies",
                "operationId": "getIrrigationEstrategiesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EntityJsonResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "jwt_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DripIrrigationModules": {
                "title": "Drip Irrigation Modules model",
                "description": "DripIrrigationModules model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the irrigation system (min: 3 characters)",
                        "type": "string",
                        "example": "module A"
                    },
                    "area": {
                        "title": "Area",
                        "description": "Area of the module ",
                        "type": "number",
                        "format": "float",
                        "example": "60"
                    },
                    "dropper_type_id": {
                        "title": "DropperTypeId",
                        "description": "The ID of the dropper type (el tipo de un gotero)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "dropper_flow": {
                        "title": "Dropper flow",
                        "description": "The flow of the dropper (flujo del gotero)",
                        "type": "number",
                        "format": "float",
                        "example": "2.0"
                    },
                    "surco_separation_id": {
                        "title": "SurcoSeparationId",
                        "description": "The ID of the surco separation (separación entre surcos)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "dropper_separation": {
                        "title": "Dropper separation",
                        "description": "The separation between droppers (separación entre goteros mts)",
                        "type": "number",
                        "format": "float",
                        "example": "0.5"
                    },
                    "irrigation_strategy_id": {
                        "title": "IrrigationStrategyId",
                        "description": "The ID of the irrigation strategy (estrategia de riego)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "irrigation_amount_minutes": {
                        "title": "IrrigationMinutes",
                        "description": "Minutes for the irrigation (minutos de riego)",
                        "type": "integer",
                        "format": "int64",
                        "example": "80"
                    },
                    "irrigation_frequency_days": {
                        "title": "IrrigationDays",
                        "description": "The amount of irrigation days per week (# días de riego a la semana)",
                        "type": "integer",
                        "format": "int64",
                        "example": "5"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DripIrrigationModules"
                }
            },
            "EstateIrrigationSystem": {
                "title": "Estate Irrigation System model",
                "description": "EstateIrrigationSystem model",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of the irrigation system (min: 3 characters)",
                        "type": "string",
                        "example": "Sistema A"
                    },
                    "minimum_tension": {
                        "title": "MinimumTension",
                        "description": "The minimum tension for the system (configuration)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "maximum_tension": {
                        "title": "MaximumTension",
                        "description": "The maximum tension for the system (configuration)",
                        "type": "integer",
                        "format": "int64",
                        "example": "70"
                    },
                    "minimum_level_water": {
                        "title": "MinimumLevelWater",
                        "description": "The minimum level of water for the system (configuration)",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "maximum_level_water": {
                        "title": "MaximumLevelWater",
                        "description": "The maximum level of water for the system (configuration)",
                        "type": "integer",
                        "format": "int64",
                        "example": "70"
                    },
                    "start_time": {
                        "title": "Start Time",
                        "description": "The initial time to start system operations (format hh:mm:ss 24h)",
                        "type": "string",
                        "format": "time",
                        "example": "07:00:00"
                    },
                    "end_time": {
                        "title": "End Time",
                        "description": "The final time of the system operations (format hh:mm:ss 24h)",
                        "type": "string",
                        "format": "time",
                        "example": "17:00:00"
                    },
                    "lamina": {
                        "title": "Lamina",
                        "description": "Lámina de agua % (número entre 0 y 1)",
                        "type": "number",
                        "format": "float",
                        "example": "1.0"
                    },
                    "irrigation_system_id": {
                        "title": "IrrigationSystemId",
                        "description": "The ID of the irrigation system ",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "estate_id": {
                        "title": "EstateId",
                        "description": "The ID of the estate (hacienda) ",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "irrigation_header_id": {
                        "title": "IrrigationHeaderId",
                        "description": "The ID of the irrigation header (cabezal de riego) ",
                        "type": "integer",
                        "format": "int64",
                        "example": "1"
                    },
                    "modules": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DripIrrigationModules"
                        },
                        "xml": {
                            "name": "DripIrrigationModules",
                            "wrapped": true
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EstateIrrigationSystem"
                }
            },
...
            "EntityJsonResource": {
                "title": "EntityJsonResource",
                "description": "A Json wrapper resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "ref": null
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EntityJsonResource"
                }
            }
        },
        "securitySchemes": {
            "jwt_key": {
                "type": "apiKey",
                "description": "Debe insertar el token generado en el proceso de autenticación",
                "name": "Authorization",
                "in": "header",
                "template": "Bearer {}"
            }
        }
    },
...
}